Origin/backend login fixes

+++ src/main/java/com/example/recipewebapp/recipedata/RecipeDataService.java
@@ -1,5 +1,7 @@
 package com.example.recipewebapp.recipedata;
 
+import com.example.recipewebapp.user.User;
+import com.example.recipewebapp.user.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +14,9 @@ import java.util.Optional;
 public class RecipeDataService {
 
     private final RecipeDataRepository recipeDataRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     @Autowired
     public RecipeDataService(RecipeDataRepository recipeDataRepository) {
@@ -29,29 +34,52 @@ public class RecipeDataService {
     }
 
 
-    public void addNewRecipe(RecipeData recipeData) {
-        Optional<RecipeData> recipeOptional = recipeDataRepository
-                .findRecipeDataByName(recipeData.getName());
-        if (recipeOptional.isPresent()) {
-            throw new IllegalStateException("name taken");
+    public void addNewRecipe(String username, RecipeData recipeData) {
+        Optional<User> userOptional = userRepository.findByUsername(username);
+        if (userOptional.isPresent()) {
+            User user = userOptional.get();
+            recipeData.setUser(user);
+            recipeDataRepository.save(recipeData);
+        } else {
+            throw new IllegalStateException("User with username " + username + " does not exist");
         }
-        recipeDataRepository.save(recipeData);
     }
 
-    public void deleteRecipe(Long recipeDataId) {
+    public void deleteRecipe(String username,Long recipeDataId) {
         boolean exists = recipeDataRepository.existsById(recipeDataId);
         if (!exists) {
             throw new IllegalStateException("recipe with id " + recipeDataId + " does not exist");
         }
+
+        RecipeData recipeData = recipeDataRepository.findById(recipeDataId)
+                .orElseThrow(() -> new IllegalStateException(
+                        "recipe with id " + recipeDataId + " does not exist"
+                ));
+
+        User currentUser = userRepository.findByUsername(username)
+                .orElseThrow(() -> new IllegalStateException("User with username " + username + " does not exist"));
+
+        if (!recipeData.getUser().equals(currentUser)) {
+            throw new IllegalStateException("User with username " + username + " is not authorized to delete this recipe");
+        }
+
         recipeDataRepository.deleteById(recipeDataId);
     }
 
     @Transactional
-    public void updateRecipe(Long recipeId, String name, String description, Integer calorieCount, String ingredients, String instructions) {
+    public void updateRecipe(String username, Long recipeId, String name, String description, Integer calorieCount, String ingredients, String instructions) {
         RecipeData recipeData = recipeDataRepository.findById(recipeId)
                 .orElseThrow(() -> new IllegalStateException(
                         "recipe with id " + recipeId + " does not exist"
                 ));
+
+        User currentUser = userRepository.findByUsername(username)
+                .orElseThrow(() -> new IllegalStateException("User with username " + username + " does not exist"));
+
+        if (!recipeData.getUser().equals(currentUser)) {
+            throw new IllegalStateException("User with username " + username + " is not authorized to edit this recipe");
+        }
+
         if (name != null &&
                 name.length() > 0 &&
                 !Objects.equals(recipeData.getName(), name)
